gdt-http now depends on gdt (not other way)
Now that gdt@v1.9.0 decoupled entirely from gdt-http, we add the
`init()` to register the gdt-http plugin with gdt's system and change
the direction of the dependency.

Users now simply need to:

```go

import (
        . "github.com/jaypipes/gdt-http"
)
```

and the gdt-http plugin self-registers with gdt, meaning the user
doesn't need to register plugins manually. All that is necessary is just
the following:

```go
func TestExample(t *testing.T) {
	fp := filepath.Join("testdata", "example.yaml")
	s, err := gdt.From(fp)

	ctx := gdt.NewContext()
	s.Run(ctx, t)
}
```

Signed-off-by: Jay Pipes <jaypipes@gmail.com>